requirements.txt
python-dotenv==1.0.0      
openai==1.2
sounddevice==0.4.6 
IPython==8.14.0
scipy==1.11.1
numpy==1.25.2